02-airflow/requirements.txt
apache-airflow-providers-google
pyarrow
openpyxl